+++ x/cronos/keeper/evm_test.go
@@ -12,7 +12,7 @@ func (suite *KeeperTestSuite) TestDeployContract() {
 	suite.SetupTest()
 	keeper := suite.app.CronosKeeper
 
-	_, err := keeper.DeployModuleCRC20(suite.ctx, "test")
+	_, err := keeper.DeployModuleCRC21(suite.ctx, "test")
 	suite.Require().NoError(err)
 }